+++ Britt2020.A.E.O/Interfaces/ResultElements/SurgeonScenarioDeviations/Id1PlusResultElement.cs
@@ -8,6 +8,6 @@
 
         IωIndexElement ωIndexElement { get; }
 
-        decimal Value { get; }
+        int Value { get; }
     }
 }
