+++ Britt2020.A.E.O/Interfaces/Results/SurgeonScenarioDeviations/Id1Plus.cs
@@ -12,7 +12,7 @@
     {
         ImmutableList<Id1PlusResultElement> Value { get; }
 
-        ImmutableList<Tuple<Organization, INullableValue<int>, INullableValue<decimal>>> GetValueForOutputContext(
+        ImmutableList<Tuple<Organization, INullableValue<int>, INullableValue<int>>> GetValueForOutputContext(
             INullableValueFactory nullableValueFactory);
     }
 }
