+++ Britt2020.A.E.O/InterfacesFactories/ResultElements/SurgeonScenarioDeviations/Id1MinusResultElementFactory.cs
@@ -8,6 +8,6 @@
         Id1MinusResultElement Create(
             IiIndexElement iIndexElement,
             IωIndexElement ωIndexElement,
-            decimal value);
+            int value);
     }
 }
