+++ Britt2020.A.E.O/InterfacesFactories/ResultElements/SurgeonScenarioDeviations/Id1PlusResultElementFactory.cs
@@ -8,6 +8,6 @@
         Id1PlusResultElement Create(
             IiIndexElement iIndexElement,
             IωIndexElement ωIndexElement,
-            decimal value);
+            int value);
     }
 }
